New python tutorial script
I really wanted to get navigation and code editing in.  The python
script gives good navigation, but the editing turned out to be mostly a
failure.  The curses builtin library has an editor, but it has a bug
that makes insert mode crash, so it has to be in overwrite mode, which
makes it really hard to edit existing sql.  It's better than nothing I
suppose.